--- .gitignore
@@ -1,12 +1,0 @@
-LinkedCodes/package-lock.json
-ter..Sanele
-node_modules
-.DS_Store
-.expo
-package-lock.json
-tsconfig.json
-cSpell.json
-package.jsonLinkedCodes/package-lock.json
-LinkedCodes/node_modules
-node_modules
-ter..Sanele
