+++ LinkedCodes/app/(tabs)/Profile/_layout.jsx
@@ -1,3 +1,13 @@
-import { Stack } from "expo-router";
+import { StatusBar } from 'expo-status-bar'
+import { Stack } from 'expo-router'
 
-export default Stack;
+const Layout = () => {
+    return (
+        <>
+            <StatusBar style="light" />
+            <Stack screenOptions={{ headerShown: false}} />
+        </>
+    )
+}
+
+export default Layout
