+++ LinkedCodes/app/(tabs)/Profile/admin/_layout.tsx
@@ -1,16 +1,17 @@
-import { View, Text } from "react-native";
 import React from "react";
 import { Redirect, Stack } from "expo-router";
 import { useUser } from "../../../../src/cxt/user";
 
 const Layout = () => {
+  const { user } = useUser();
 
-    const { user } = useUser();
+  if (!user?.admin) return <Redirect href="/(tabs)/Profile" />
 
-    if (!user?.admin) return <Redirect href="/(tabs)/Profile" />
-
-
-  return <Stack />;
+  return (
+    <>
+      <Stack screenOptions={{ headerShown: true }} />
+    </>
+  );
 };
 
 export default Layout;
