--- LinkedCodes/app/(userTabs)/home/notifications.jsx
@@ -1,14 +1,0 @@
-import { StyleSheet, Text, View } from 'react-native'
-import React from 'react'
-
-const Notification = () => {
-  return (
-    <View>
-      <Text>Notification</Text>
-    </View>
-  )
-}
-
-export default Notification
-
-const styles = StyleSheet.create({})
