--- LinkedCodes/src/components/try.jsx
@@ -1,12 +1,0 @@
-import { View, Text } from 'react-native'
-import React from 'react'
-
-const Try = () => {
-  return (
-    <View>
-      <Text>try</Text>
-    </View>
-  )
-}
-
-export default Try
