--- LinkedCodes/src/profile/firebase.jsx
@@ -1,2 +1,0 @@
-import { auth, db, storage } from "../../firebase"
-export { auth, db, storage };
